requirements.txt
discord.py==2.6.4
python-dotenv==1.2.1
cryptography==46.0.5
pymongo==4.16.0
tzdata==2025.3
aiohttp==3.11.14
pytest
pytest-asyncio
pytest-cov
